backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "alex-database",
    "bedrock-agentcore>=1.0.3",
    "bedrock-agentcore-starter-toolkit>=0.1.26",
    "boto3>=1.40.29",
    "langfuse>=3.3.4",
    "nest-asyncio>=1.6.0",
    "openai-agents>=0.3.0",
    "playwright>=1.55.0",
    "pydantic-ai>=1.0.6",
    "python-dotenv>=1.1.1",
    "strands-agents>=1.13.0",
    "strands-agents-tools>=0.2.12",
]

[tool.uv.workspace]
members = [
    "database",
    "api",
    "scheduler",
]

[tool.uv.sources]
alex-database = { workspace = true }
